下单使用小数修改成功

+++ app/src/main/java/com/runwise/supply/orderpage/OneKeyActivityV2.java
@@ -92,7 +92,7 @@ public class OneKeyActivityV2 extends ProductActivityV2 {
                     Toast.makeText(this,"小主，暂时不用采购哦~",Toast.LENGTH_LONG).show();
                 }else{
                     for(ProductData.ListBean pBean: data.getList()){
-                        mMapCount.put(pBean,pBean.getPresetQty());
+                        mMapCount.put(pBean,(double)pBean.getPresetQty());
                     }
                 }
                 requestCategory();
